Moved error component into commons
