+++ frontend/typescript/components/app/app.component.ts
@@ -9,8 +9,8 @@ import {RAMNavService} from '../../services/ram-nav.service';
 import {RAMConstantsService} from '../../services/ram-constants.service';
 
 import {MyIdentityComponent} from '../commons/my-identity/my-identity.component';
-import {ErrorComponent} from '../error/error.component';
-import {ErrorService} from '../error/error.service';
+import {ErrorComponent} from '../commons/error/error.component';
+import {ErrorService} from '../commons/error/error.service';
 
 @Component({
     selector: 'ram-app',
