+++ frontend/typescript/components/commons/error/error.component.ts
